move FreDF tests to existing loss test file
