+++ tests/test_losses/test_pytorch.py
@@ -6,6 +6,7 @@ from neuralforecast.losses.pytorch import (
     MAE,
     PMM,
     DistributionLoss,
+    FreDF,
     HuberIQLoss,
     IQLoss,
     MQLoss,
@@ -160,3 +161,47 @@ def test_duplicate_level_and_quantiles_dedup():
         check = MQLoss(level=[80, 90])
         assert len(w) == 0
     assert len(check.quantiles) == 5
+
+
+def test_fredf_alpha0_equals_mse():
+    """When alpha=0, FreDF should reduce to pure time-domain MSE."""
+    loss_fn = FreDF(alpha=0.0)
+
+    y     = torch.tensor([[[1.0], [2.0], [3.0]]])  # [1, 3, 1]
+    y_hat = torch.tensor([[[2.0], [2.0], [2.0]]])
+
+    loss = loss_fn(y, y_hat)
+
+    # MSE = mean((1-2)^2, (2-2)^2, (3-2)^2) = mean(1, 0, 1) = 2/3
+    expected = torch.tensor(2.0 / 3.0)
+    assert torch.isclose(loss, expected, atol=1e-5), \
+        f"Expected {expected.item():.6f}, got {loss.item():.6f}"
+
+
+def test_fredf_alpha1_equals_freq_mae():
+    """When alpha=1, FreDF should reduce to pure frequency-domain MAE."""
+    loss_fn = FreDF(alpha=1.0)
+
+    y     = torch.tensor([[[1.0], [2.0], [3.0]]])  # [1, 3, 1]
+    y_hat = torch.tensor([[[2.0], [2.0], [2.0]]])
+
+    loss = loss_fn(y, y_hat)
+
+    # rfft(y)     = [6+0j, -1.5+0.866j]
+    # rfft(y_hat) = [6+0j, 0+0j]
+    # |diff|      = [0.0, 1.732]
+    # MAE         = mean(0.0, 1.732) = 0.866
+    expected = torch.tensor(0.866025)
+    assert torch.isclose(loss, expected, atol=1e-4), \
+        f"Expected {expected.item():.6f}, got {loss.item():.6f}"
+
+
+def test_fredf_perfect_forecast_zero_loss():
+    """Perfect forecast should give zero loss for any alpha."""
+    y = torch.tensor([[[1.0], [2.0], [3.0]]])
+
+    for alpha in [0.0, 0.5, 1.0]:
+        loss_fn = FreDF(alpha=alpha)
+        loss = loss_fn(y, y)
+        assert torch.isclose(loss, torch.tensor(0.0), atol=1e-6), \
+            f"alpha={alpha}: expected 0.0, got {loss.item():.6f}"
